Modify the speed and rpm calculation part
